Neue Klasse DistanceChecker
